Misc refactors
- Turn hoaId to long wherever it is an integer
- Standardize error message types

+++ voting-microservice/src/main/java/voting/domain/Election.java
@@ -107,7 +107,7 @@ public abstract class Election {
         this.electionId = electionId;
     }
 
-    public void setHoaId(int hoaId) {
+    public void setHoaId(long hoaId) {
         this.hoaId = hoaId;
     }
